issue: 226 Expose CP for external HomeStore client
